% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imrsts.R
\name{doGenerate}
\alias{doGenerate}
\title{Generate an html output consists of plots and table.}
\usage{
doGenerate(mySTS, ts = NULL)
}
\arguments{
\item{mySTS}{text.}

\item{ts}{timestamp. Must be in "YYMMDD.HHMMSS" or "YYMMDD" character format. Default is NULL.}
}
\value{
If the input valid and the time series contains a build fingerprint, a 
  list will be returned. Otherwise it's NULL.
}
\description{
\code{doGenerate} generates a html page (using knitr), which consists of plots 
and tables and also returns the metadata an survey time series specified in its argument.
}
\examples{
\dontrun{
doGenerate("Barents Sea Northeast Arctic cod bottom trawl index in winter")
}

}
