% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGen.R
\name{doGenReport}
\alias{doGenReport}
\title{Automatic report generation of plots and table from a specific directory.}
\usage{
doGenReport(srcdir, verbose = FALSE)
}
\arguments{
\item{srcdir}{directory containing the files.}

\item{verbose}{whether to print all included files in the R console. Default to FALSE.}
}
\description{
\code{doGenReport} generates a html page using knitr from all CSV, PNG, and TXT files inside a directory.
}
\examples{
\dontrun{
doGenReport("./projects")
}

}
