% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{getErrorInfo}
\alias{getErrorInfo}
\title{Get error information of a specific item in the matrix.}
\usage{
getErrorInfo(path, timestamp)
}
\arguments{
\item{path}{The path to the item.}

\item{timestamp}{The timestamp of the iter.}
}
\description{
\code{loadTimeSeriesStatus} returns the status of the timeseries processing.
}
