% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getREDUSRootDir}
\alias{getREDUSRootDir}
\title{Get REDUS work directory (as defined in the option file)}
\usage{
getREDUSRootDir()
}
\description{
Get REDUS work directory (as defined in the option file)
}
