% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imrsts.R
\name{getSTSList}
\alias{getSTSList}
\title{Load the available STS names from IMR NMD.}
\usage{
getSTSList()
}
\value{
A list contains all the names of the available STS will be returned.
}
\description{
\code{getSTSList} returns an list of names of the available survey time series.
}
