% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{loadTimeSeriesStatus}
\alias{loadTimeSeriesStatus}
\title{Load time series status data.}
\usage{
loadTimeSeriesStatus(type = "REDUS")
}
\arguments{
\item{type}{The type of data source requested. Defaults to "REDUS".}
}
\description{
\code{loadTimeSeriesStatus} returns the status of the timeseries processing.
}
\examples{
loadTimeSeriesStatus()
}
