% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{preprocess}
\alias{preprocess}
\title{Start the pre-processing based on the given configuration file.}
\usage{
preprocess(configFile)
}
\arguments{
\item{configFile}{The path to the configuration file.}
}
\description{
\code{preprocess} returns TRUE if succeeded or FALSE if failed.
}
