% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imrsts.R
\name{processLenGrpHeader}
\alias{processLenGrpHeader}
\title{Process a length grup header.}
\usage{
processLenGrpHeader(tblResult, target = "LenGrp", prefix = "",
  rangeSeparator = "-", plusAgeChar = "+", sep = " ")
}
\arguments{
\item{tblResult}{REDUS STS data table.}

\item{target}{Column name of the length groups category.}

\item{prefix}{Prefix for the headers.}

\item{rangeSeparator}{Range separator character.}

\item{plusAgeChar}{Plus age character.}

\item{sep}{Separator character.}
}
\value{
On a valid input a character object (list) will  be returned.
}
\description{
\code{processLenGrpHeader} returns a length group header translation.
}
\details{
For example an original length group header of \code{c('5', '10', '15')} is transformed
into \code{c('5'='5-9', '10'='10-15', '15'='15+')}.
}
\examples{
stsData <- loadSTS("Barents Sea Northeast Arctic cod bottom trawl index in winter")
headerTranslation <- processLenGrpHeader(stsData$data2)

}
