% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxProcess.R
\name{processRstoxSTS}
\alias{processRstoxSTS}
\title{Process survey time series using Rstox.}
\usage{
processRstoxSTS(masterScript)
}
\arguments{
\item{masterScript}{Location of the REDUS master script configuration.}
}
\description{
\code{processRstoxSTS} returns a detailed status of the actions performed in survey time series process.
}
