% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{read.surveys.with.negative}
\alias{read.surveys.with.negative}
\title{Function to read ices survey format. Copied from:
https://github.com/fishfollower/SAM/blob/master/stockassessment/R/reading.R
The only difference is that this function doesn't remove negative numbers}
\usage{
read.surveys.with.negative(filen)
}
\arguments{
\item{filen}{the file}
}
\description{
Function to read ices survey format. Copied from:
https://github.com/fishfollower/SAM/blob/master/stockassessment/R/reading.R
The only difference is that this function doesn't remove negative numbers
}
