% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imrsts.R
\name{stsPlot}
\alias{stsPlot}
\title{Generate an STS plot using ggplot2.}
\usage{
stsPlot(stsData, type = "Abundance")
}
\arguments{
\item{stsData}{REDUS STS data from \code{loadSTS}.}

\item{type}{text, valid values are \code{"Abundance"} and \code{"Weight"}.}
}
\value{
On a valid input a ggplot2 object will be returned.
}
\description{
\code{stsPlot} returns an STS plot based on its parameters.
}
\examples{
\dontrun{
stsPlot(loadSTS("Barents Sea Northeast Arctic cod bottom trawl index in winter"), "Abundance")
}

}
