% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imrsts.R
\name{stsTable}
\alias{stsTable}
\title{Generate an STS data table.}
\usage{
stsTable(stsData, type = "Abundance", raw = FALSE)
}
\arguments{
\item{stsData}{REDUS STS data from \code{loadSTS}.}

\item{type}{text, valid values are \code{"Abundance"} and \code{"Weight"}.}

\item{raw}{boolean}
}
\description{
\code{stsTable} prints out a HTML table based on its parameters.
}
\examples{
stsTable(loadSTS("Barents Sea Northeast Arctic cod bottom trawl index in winter"), "Abundance")

}
