% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{updateICESSurveyFile}
\alias{updateICESSurveyFile}
\title{Update or add a survey index abundance in an ICES format file.}
\usage{
updateICESSurveyFile(surveyFile, type, text, srcHead, stsName,
  stsDate = NULL, useSourceAge = FALSE, useSourceYear = FALSE,
  target)
}
\arguments{
\item{surveyFile}{The path to the survey file.}

\item{type}{Data type from source to put into the file. Either "data" or "data_var" (covariance).}

\item{text}{If not NULL then text data in this variable will be put into the survey file.}

\item{srcHead}{Survey title to be (partially) matched.}

\item{stsName}{Survey time series name that will served as the source.}

\item{stsDate}{The timestamp of the survey time series source.}

\item{useSourceAge}{The timestamp of the iter.}

\item{useSourceYear}{The timestamp of the iter.}

\item{target}{The timestamp of the iter.}
}
\description{
\code{updateICESSurveyFile} returns TRUE if succeeded or FALSE if failed.
}
