% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{updateICESSurveyFile}
\alias{updateICESSurveyFile}
\title{Update or add a survey index abundance in an ICES format file.}
\usage{
updateICESSurveyFile(
  surveyFile,
  text,
  srcHead,
  stsName,
  stsDate = NULL,
  useSourceAge = FALSE,
  useSourceYear = FALSE,
  target
)
}
\arguments{
\item{surveyFile}{The path to the survey file.}

\item{text}{If not NULL then text data in this variable will be put into the survey file.}

\item{srcHead}{Survey title to be (partially) matched.}

\item{stsName}{Survey time series name that will served as the source.}

\item{stsDate}{The timestamp of the survey time series source.}

\item{useSourceAge}{Use similar age structure to the existing/source survey file.}

\item{useSourceYear}{Use similar year structure to the existing/source survey file.}

\item{target}{Target mode (local or remote).}
}
\description{
\code{updateICESSurveyFile} returns TRUE if succeeded or FALSE if failed.
}
